+++ Code/Renderer/rendererGlobals.h
@@ -179,10 +179,10 @@ uint32_t HostDataBufferObject<T>::count()
 }
 
 template <typename T>
-HostDataBufferObject<T> createDataBuffer(RendererContext* h, uint32_t size, VkBufferUsageFlags usage)
+HostDataBufferObject<T> createDataBuffer(RendererContext* h, size_t size, VkBufferUsageFlags usage)
 {
     HostDataBufferObject<T> hostDataBuffer{};
-    hostDataBuffer.buffer.size = sizeof(T) * size;
+    hostDataBuffer.buffer.size = (uint32_t)(sizeof(T) * size);
     hostDataBuffer.buffer.mapped = BufferUtilities::createHostMappedBuffer(
         h->allocator, hostDataBuffer.buffer.size, usage,
         &hostDataBuffer.allocation,
